--- a/crates/rune/src/runtime/vm.rs
+++ b/crates/rune/src/runtime/vm.rs
@@ -927,15 +927,17 @@ impl Vm {
             }
             target => {
                 let index = index.clone();
-                return Ok(match self
-                    .call_field_fn(Protocol::GET, target, index.hash(), ())?
-                    {
+                return Ok(
+                    match self.call_field_fn(Protocol::GET, target, index.hash(), ())? {
                         CallResult::Ok(()) => return Ok(CallResult::Ok(self.stack.pop()?)),
-                        CallResult::Unsupported(target) => match self.call_instance_fn(target, Protocol::FALLBACK_GET, (index,))? {
-                            CallResult::Ok(()) => CallResult::Ok(self.stack.pop()?),
-                            CallResult::Unsupported(target) => CallResult::Unsupported(target),
-                        },
-                    });
+                        CallResult::Unsupported(target) => {
+                            match self.call_instance_fn(target, Protocol::FALLBACK_GET, (index,))? {
+                                CallResult::Ok(()) => CallResult::Ok(self.stack.pop()?),
+                                CallResult::Unsupported(target) => CallResult::Unsupported(target),
+                            }
+                        }
+                    },
+                );
             }
         }
 
@@ -988,15 +990,23 @@ impl Vm {
             }
             target => {
                 let index = field.clone();
-                return Ok(match self
-                    .call_field_fn(Protocol::SET, target, index.hash(), (value.clone(),))?
-                    {
+                return Ok(
+                    match self.call_field_fn(
+                        Protocol::SET,
+                        target,
+                        index.hash(),
+                        (value.clone(),),
+                    )? {
                         CallResult::Ok(()) => {
                             self.stack.pop()?;
                             CallResult::Ok(())
                         }
                         CallResult::Unsupported(target) => {
-                            match self.call_instance_fn(target, Protocol::FALLBACK_SET, (index, value))? {
+                            match self.call_instance_fn(
+                                target,
+                                Protocol::FALLBACK_SET,
+                                (index, value),
+                            )? {
                                 CallResult::Ok(()) => {
                                     self.stack.pop()?;
                                     CallResult::Ok(())
@@ -1004,7 +1014,8 @@ impl Vm {
                                 CallResult::Unsupported(target) => CallResult::Unsupported(target),
                             }
                         }
-                    });
+                    },
+                );
             }
         })
     }